+++ federatedscope/core/data/__init__.py
@@ -1,4 +1,4 @@
 from federatedscope.core.data.base_data import StandaloneDataDict, ClientData
-from federatedscope.core.data.data_translator import BaseDataTranslator
+from federatedscope.core.data.base_translator import BaseDataTranslator
 
 __all__ = ['StandaloneDataDict', 'ClientData', 'BaseDataTranslator']
